+++ src/app/workflow.rs
@@ -1,10 +1,7 @@
 use crate::{
-    domain::{
-        article::{get_article_content_with_client, store_article_content, ArticleContent},
-        feed::{search_feeds, Feed, FeedQuery},
-        rss::{get_article_links_from_feed, search_backlog_article_links, store_article_links},
-    },
+    core::feed::{search_feeds, FeedQuery},
     infra::api::{firecrawl::FirecrawlClient, http::HttpClient},
+    task::{process_collect_article_links, process_collect_articles},
 };
 use anyhow::{Context, Result};
 use sqlx::PgPool;
@@ -62,629 +59,15 @@ pub async fn execute_rss_workflow<H: HttpClient, F: FirecrawlClient>(
     Ok(())
 }
 
-/// RSSフィードからリンクを収集してDBに保存する
-async fn process_collect_article_links<H: HttpClient>(
-    client: &H,
-    feeds: &[Feed],
-    pool: &PgPool,
-) -> Result<()> {
-    println!("--- RSSフィードからリンク取得開始 ---");
-
-    for feed in feeds {
-        println!("フィード処理中: {}", feed);
-
-        match get_article_links_from_feed(client, feed).await {
-            Ok(article_links) => {
-                println!("  {}件のリンクを抽出", article_links.len());
-
-                match store_article_links(&article_links, pool).await {
-                    Ok(_) => {
-                        println!("  DB保存完了: {}件処理", article_links.len());
-                    }
-                    Err(e) => {
-                        eprintln!("  DB保存エラー: {}", e);
-                    }
-                }
-            }
-            Err(e) => {
-                eprintln!("  フィード取得エラー: {}", e);
-            }
-        }
-    }
-
-    println!("--- RSSフィードからリンク取得完了 ---");
-    Ok(())
-}
-
-/// バックログ対象リンクから処理待ちの記事を収集してDBに保存する
-async fn process_collect_articles<F: FirecrawlClient>(
-    firecrawl_client: &F,
-    pool: &PgPool,
-) -> Result<()> {
-    println!("--- 記事内容取得開始 ---");
-    // 未処理のリンクを取得（articleテーブルに存在しないarticle_linkを取得）
-    let unprocessed_links = search_backlog_article_links(pool).await?;
-    println!("未処理リンク数: {}件", unprocessed_links.len());
-
-    for article_link in unprocessed_links {
-        println!("記事処理中: {}", article_link.url);
-
-        let article_result =
-            get_article_content_with_client(&article_link.url, firecrawl_client).await;
-
-        match article_result {
-            Ok(article) => match store_article_content(&article, pool).await {
-                Ok(_) => {
-                    println!("  記事保存完了");
-                }
-                Err(e) => {
-                    eprintln!("  記事保存エラー: {}", e);
-                }
-            },
-            Err(e) => {
-                eprintln!("  記事取得エラー: {}", e);
-
-                // エラーが発生した場合も、status_codeを記録してスキップ
-                let error_article = ArticleContent {
-                    url: article_link.url,
-                    timestamp: chrono::Utc::now(),
-                    status_code: 500, // エラー用のステータスコード
-                    content: format!("取得エラー: {}", e),
-                };
-
-                if let Err(store_err) = store_article_content(&error_article, pool).await {
-                    eprintln!("  エラー記事の保存に失敗: {}", store_err);
-                }
-            }
-        }
-    }
-
-    println!("--- 記事内容取得完了 ---");
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use sqlx::PgPool;
 
-    mod unit {
-        use super::*;
-        use crate::infra::api::firecrawl::MockFirecrawlClient;
-
-        #[sqlx::test(fixtures("../../fixtures/workflow.sql"))]
-        async fn test_process_collect_articles(pool: PgPool) -> Result<(), anyhow::Error> {
-            // fixtureから6件の未処理RSSリンクと3件の処理済み記事が読み込まれる（archiveも再処理される）
-
-            // 全URL成功のモッククライアントを設定（基本テスト用）
-            let mock_client = MockFirecrawlClient::new_success("基本テスト記事の内容です");
-            // 記事取得を実行（未処理の6件が処理される）
-            let result = process_collect_articles(&mock_client, &pool).await;
-            assert!(
-                result.is_ok(),
-                "記事取得処理が失敗しました: {:?}",
-                result.err()
-            );
-            // 全記事数確認（既存3件 + 新規3件 + 更新3件 = 9件、実際は再処理により既存が更新されて8件）
-            let total_articles = sqlx::query_scalar!("SELECT COUNT(*) FROM articles")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                total_articles.unwrap_or(0),
-                8,
-                "総記事数が期待値と異なります"
-            );
-            // 成功記事数確認（全て成功で処理されるため8件）
-            let new_success_articles =
-                sqlx::query_scalar!("SELECT COUNT(*) FROM articles WHERE status_code = 200")
-                    .fetch_one(&pool)
-                    .await?;
-            assert_eq!(
-                new_success_articles.unwrap_or(0),
-                8,
-                "成功記事数が期待値と異なります"
-            );
-            // エラー記事数の確認（全て成功処理されるため0件）
-            let error_articles =
-                sqlx::query_scalar!("SELECT COUNT(*) FROM articles WHERE status_code = 500")
-                    .fetch_one(&pool)
-                    .await?;
-            assert_eq!(
-                error_articles.unwrap_or(0),
-                0,
-                "エラー記事数が期待値と異なります"
-            );
-            // 特定記事の内容確認
-            let article_content: String = sqlx::query_scalar!(
-                "SELECT content FROM articles WHERE url = $1",
-                "https://news.example.com/article1"
-            )
-            .fetch_one(&pool)
-            .await?;
-            assert!(
-                article_content.contains("基本テスト記事の内容です"),
-                "記事内容が正しく保存されていません"
-            );
-
-            println!("✅ 基本workflow統合テスト完了: 6件の記事を処理しました");
-            Ok(())
-        }
-
-        #[sqlx::test(fixtures("../../fixtures/workflow_mixed.sql"))]
-        async fn test_process_collect_articles_mixed(pool: PgPool) -> Result<(), anyhow::Error> {
-            // fixtureから11件の未処理RSSリンクと2件の処理済み記事が読み込まれる（エラー記事も再処理）
-
-            // 全URL成功のモッククライアントを設定（混在テスト用）
-            let mock_client = MockFirecrawlClient::new_success("混在テスト記事の内容です");
-            // 記事取得を実行（未処理の11件が処理される）
-            let result = process_collect_articles(&mock_client, &pool).await;
-            assert!(
-                result.is_ok(),
-                "混在シナリオの処理が失敗しました: {:?}",
-                result.err()
-            );
-            // 全記事数確認（既存2件 + 新規10件 = 12件）
-            let total_articles = sqlx::query_scalar!("SELECT COUNT(*) FROM articles")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                total_articles.unwrap_or(0),
-                12,
-                "総記事数が期待値と異なります"
-            );
-            // 成功記事数確認（全て成功で処理されるため12件）
-            let success_articles =
-                sqlx::query_scalar!("SELECT COUNT(*) FROM articles WHERE status_code = 200")
-                    .fetch_one(&pool)
-                    .await?;
-            assert_eq!(
-                success_articles.unwrap_or(0),
-                12,
-                "成功記事数が期待値と異なります"
-            );
-            // エラー記事数確認（全て成功処理されるため0件）
-            let error_articles =
-                sqlx::query_scalar!("SELECT COUNT(*) FROM articles WHERE status_code = 500")
-                    .fetch_one(&pool)
-                    .await?;
-            assert_eq!(
-                error_articles.unwrap_or(0),
-                0,
-                "エラー記事数が期待値と異なります"
-            );
-            // 成功記事の内容確認（全て成功するのでいずれかの記事を確認）
-            let success_content: String = sqlx::query_scalar!(
-                "SELECT content FROM articles WHERE url = $1",
-                "https://success.example.com/news1"
-            )
-            .fetch_one(&pool)
-            .await?;
-            assert!(
-                success_content.contains("混在テスト記事の内容です"),
-                "成功記事の内容が正しくありません"
-            );
-
-            println!("✅ 混在シナリオworkflow統合テスト完了: 11件すべて成功処理しました");
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_process_collect_article_links_success(
-            pool: PgPool,
-        ) -> Result<(), anyhow::Error> {
-            use crate::domain::feed::Feed;
-            use crate::infra::api::http::MockHttpClient;
-
-            // テスト用フィードを準備（異なるURLで3つのフィード）
-            let test_feeds = vec![
-                Feed {
-                    group: "news".to_string(),
-                    name: "tech_news".to_string(),
-                    rss_link: "https://technews.example.com/rss.xml".to_string(),
-                },
-                Feed {
-                    group: "blog".to_string(),
-                    name: "dev_blog".to_string(),
-                    rss_link: "https://devblog.example.com/feed.xml".to_string(),
-                },
-                Feed {
-                    group: "updates".to_string(),
-                    name: "product_updates".to_string(),
-                    rss_link: "https://updates.example.com/rss".to_string(),
-                },
-            ];
-
-            // MockHttpClientで成功レスポンスを設定
-            let mock_client = MockHttpClient::new_success();
-
-            // 処理前のarticle_links件数を確認
-            let initial_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                initial_count.unwrap_or(0),
-                0,
-                "初期状態でarticle_linksが空ではありません"
-            );
-
-            // process_collect_article_linksを実行
-            let result = process_collect_article_links(&mock_client, &test_feeds, &pool).await;
-            assert!(
-                result.is_ok(),
-                "RSS収集処理が失敗しました: {:?}",
-                result.err()
-            );
-
-            // 処理後のarticle_links件数を確認（3フィード × 3記事 = 9件）
-            let final_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                final_count.unwrap_or(0),
-                9,
-                "期待されるarticle_links件数と異なります"
-            );
-
-            // 各フィードから生成されたリンクの形式を検証
-            use crate::infra::compute::generate_mock_rss_id;
-
-            for feed in &test_feeds {
-                let hash = generate_mock_rss_id(&feed.rss_link);
-
-                // 各フィードから3件のリンクが生成されていることを確認
-                let feed_link_count = sqlx::query_scalar!(
-                    "SELECT COUNT(*) FROM article_links WHERE url LIKE $1",
-                    format!("https://{}.example.com/%", hash)
-                )
-                .fetch_one(&pool)
-                .await?;
-
-                assert_eq!(
-                    feed_link_count.unwrap_or(0),
-                    3,
-                    "フィード {} から3件のリンクが生成されるべきです",
-                    feed
-                );
-
-                // タイトルの形式検証（{hash}:title:X の形式）
-                for article_num in 1..=3 {
-                    let expected_title = format!("{}:title:{}", hash, article_num);
-                    let expected_link = format!("https://{}.example.com/{}", hash, article_num);
-
-                    let title_exists = sqlx::query_scalar!(
-                        "SELECT COUNT(*) FROM article_links WHERE title = $1 AND url = $2",
-                        expected_title,
-                        expected_link
-                    )
-                    .fetch_one(&pool)
-                    .await?;
-
-                    assert_eq!(
-                        title_exists.unwrap_or(0),
-                        1,
-                        "期待されるタイトル '{}' とリンク '{}' の組み合わせが見つかりません",
-                        expected_title,
-                        expected_link
-                    );
-                }
-            }
-
-            // 動的生成された日付が適切な範囲に設定されていることを確認（3日前～今日）
-            let date_count = sqlx::query_scalar!(
-                "SELECT COUNT(*) FROM article_links WHERE pub_date BETWEEN $1 AND $2",
-                chrono::Utc::now() - chrono::Duration::days(3),
-                chrono::Utc::now() + chrono::Duration::hours(1)
-            )
-            .fetch_one(&pool)
-            .await?;
-
-            assert_eq!(
-                date_count.unwrap_or(0),
-                9,
-                "すべてのリンクの日付が動的生成範囲（3日以内）にありません"
-            );
-
-            println!("✅ RSS収集基本テスト完了");
-            println!("  処理されたフィード数: {}", test_feeds.len());
-            println!("  保存されたリンク数: {}", final_count.unwrap_or(0));
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_process_collect_article_links_with_errors(
-            pool: PgPool,
-        ) -> Result<(), anyhow::Error> {
-            use crate::domain::feed::Feed;
-            use crate::infra::api::http::MockHttpClient;
-
-            // 成功フィード1つ + エラーフィード2つを準備
-            let test_feeds = vec![
-                Feed {
-                    group: "success".to_string(),
-                    name: "working_feed".to_string(),
-                    rss_link: "https://working.example.com/rss.xml".to_string(),
-                },
-                Feed {
-                    group: "error1".to_string(),
-                    name: "timeout_feed".to_string(),
-                    rss_link: "https://timeout.example.com/rss.xml".to_string(),
-                },
-                Feed {
-                    group: "error2".to_string(),
-                    name: "server_error_feed".to_string(),
-                    rss_link: "https://servererror.example.com/rss.xml".to_string(),
-                },
-            ];
-
-            // 成功クライアントで正常フィードを処理
-            let success_client = MockHttpClient::new_success();
-
-            // process_collect_article_linksは内部的にはエラーを握り潰して継続処理するため、
-            // 個別にテストする必要がある
-
-            // 1. 成功フィードのテスト
-            let success_feeds = vec![test_feeds[0].clone()];
-            let result =
-                process_collect_article_links(&success_client, &success_feeds, &pool).await;
-            assert!(result.is_ok(), "成功フィードの処理が失敗しました");
-
-            // 成功フィードからの3件のリンクが保存されることを確認
-            let success_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                success_count.unwrap_or(0),
-                3,
-                "成功フィードから3件のリンクが保存されるべきです"
-            );
-
-            // 2. エラークライアントで全フィードを処理
-            let error_client = MockHttpClient::new_error("接続タイムアウト");
-
-            // エラークライアントでも処理自体は成功する（内部でエラーハンドリング）
-            let all_result = process_collect_article_links(&error_client, &test_feeds, &pool).await;
-            assert!(
-                all_result.is_ok(),
-                "エラーハンドリングが正しく動作していません"
-            );
-
-            // エラーフィードからは新たなリンクが追加されないことを確認
-            let final_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                final_count.unwrap_or(0),
-                3,
-                "エラーフィードから新たなリンクが追加されるべきではありません"
-            );
-
-            // 3. 成功・エラー混在での処理確認
-            // 新しいテーブル状態でテスト
-            sqlx::query!("DELETE FROM article_links")
-                .execute(&pool)
-                .await?;
-
-            // 混在処理では各フィードが個別に処理される
-            // この関数は現在の実装ではクライアント固定なので、実際の混在テストは困難
-            // その代わりに、成功ケースが正しく処理されることを再確認
-            let final_result =
-                process_collect_article_links(&success_client, &success_feeds, &pool).await;
-            assert!(
-                final_result.is_ok(),
-                "最終的な成功フィード処理が失敗しました"
-            );
-
-            let final_success_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                final_success_count.unwrap_or(0),
-                3,
-                "最終的な成功フィード処理結果が不正です"
-            );
-
-            println!("✅ RSSエラーハンドリングテスト完了");
-            println!("  エラーがあっても処理が継続されることを確認");
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_process_collect_article_links_duplicate_handling(
-            pool: PgPool,
-        ) -> Result<(), anyhow::Error> {
-            use crate::domain::feed::Feed;
-            use crate::infra::api::http::MockHttpClient;
-
-            // 同一URLを持つ複数のフィードを準備（重複リンクを意図的に生成）
-            let same_rss_url = "https://shared.example.com/common.xml";
-            let duplicate_feeds = vec![
-                Feed {
-                    group: "group1".to_string(),
-                    name: "shared_feed_1".to_string(),
-                    rss_link: same_rss_url.to_string(),
-                },
-                Feed {
-                    group: "group2".to_string(),
-                    name: "shared_feed_2".to_string(),
-                    rss_link: same_rss_url.to_string(),
-                },
-                Feed {
-                    group: "group3".to_string(),
-                    name: "shared_feed_3".to_string(),
-                    rss_link: same_rss_url.to_string(),
-                },
-            ];
-
-            // MockHttpClientで成功レスポンスを設定
-            let mock_client = MockHttpClient::new_success();
-
-            // 初期状態の確認
-            let initial_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                initial_count.unwrap_or(0),
-                0,
-                "初期状態でarticle_linksが空ではありません"
-            );
-
-            // 1回目の実行：最初のフィードを処理
-            let first_feed = vec![duplicate_feeds[0].clone()];
-            let result1 = process_collect_article_links(&mock_client, &first_feed, &pool).await;
-            assert!(result1.is_ok(), "1回目のRSS収集処理が失敗しました");
-
-            // 1回目実行後の件数確認（3件のリンクが挿入されるはず）
-            let after_first_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                after_first_count.unwrap_or(0),
-                3,
-                "1回目実行後に3件のリンクが保存されるべきです"
-            );
-
-            // 1回目実行後の日付を記録（更新確認のため）
-            let first_pub_dates: Vec<chrono::DateTime<chrono::Utc>> =
-                sqlx::query_scalar!("SELECT pub_date FROM article_links ORDER BY url")
-                    .fetch_all(&pool)
-                    .await?;
-            assert_eq!(
-                first_pub_dates.len(),
-                3,
-                "1回目実行後に3件の日付が記録されるべきです"
-            );
-
-            // 少し待機して、動的日付生成で異なる時刻になることを確保
-            tokio::time::sleep(tokio::time::Duration::from_millis(1000)).await;
-
-            // 2回目の実行：同一URLのフィードを再度処理（重複発生）
-            let second_feed = vec![duplicate_feeds[1].clone()];
-            let result2 = process_collect_article_links(&mock_client, &second_feed, &pool).await;
-            assert!(result2.is_ok(), "2回目のRSS収集処理が失敗しました");
-
-            // 2回目実行後の件数確認（重複により件数は変わらず3件のまま）
-            let after_second_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                after_second_count.unwrap_or(0),
-                3,
-                "2回目実行後も3件のまま（重複時は上書き更新）であるべきです"
-            );
-
-            // 2回目実行後の日付を取得して更新状況を確認
-            let second_pub_dates: Vec<chrono::DateTime<chrono::Utc>> =
-                sqlx::query_scalar!("SELECT pub_date FROM article_links ORDER BY url")
-                    .fetch_all(&pool)
-                    .await?;
-            assert_eq!(
-                second_pub_dates.len(),
-                3,
-                "2回目実行後も3件の日付が記録されているべきです"
-            );
-
-            // 重複リンクの場合、日付は更新される（ON CONFLICT DO UPDATE）
-            for (i, (first_date, second_date)) in first_pub_dates
-                .iter()
-                .zip(second_pub_dates.iter())
-                .enumerate()
-            {
-                assert_ne!(
-                    first_date,
-                    second_date,
-                    "記事{}の日付が更新されませんでした（重複時は新しい日付で更新されるべき）: {} == {}",
-                    i + 1,
-                    first_date,
-                    second_date
-                );
-                assert!(
-                    second_date >= first_date,
-                    "記事{}の日付が過去に戻りました（新しい日付のほうが新しいべき）: {} < {}",
-                    i + 1,
-                    second_date,
-                    first_date
-                );
-            }
-
-            // 3回目の実行：全ての重複フィードを一度に処理
-            let all_result =
-                process_collect_article_links(&mock_client, &duplicate_feeds, &pool).await;
-            assert!(all_result.is_ok(), "全重複フィードの処理が失敗しました");
-
-            // 最終的な件数確認（依然として3件のまま）
-            let final_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                final_count.unwrap_or(0),
-                3,
-                "最終的にも3件のまま（すべての重複が上書き更新）であるべきです"
-            );
-
-            // 保存されたリンクの内容確認
-            use crate::infra::compute::generate_mock_rss_id;
-            let expected_hash = generate_mock_rss_id(same_rss_url);
-
-            for article_num in 1..=3 {
-                let expected_title = format!("{}:title:{}", expected_hash, article_num);
-                let expected_link =
-                    format!("https://{}.example.com/{}", expected_hash, article_num);
-
-                let link_exists = sqlx::query_scalar!(
-                    "SELECT COUNT(*) FROM article_links WHERE title = $1 AND url = $2",
-                    expected_title,
-                    expected_link
-                )
-                .fetch_one(&pool)
-                .await?;
-
-                assert_eq!(
-                    link_exists.unwrap_or(0),
-                    1,
-                    "期待されるリンク '{}' が1件だけ存在すべきです（重複なし）",
-                    expected_link
-                );
-            }
-
-            // 異なるURLのフィードを追加して、重複処理が新規リンクをブロックしないことを確認
-            let unique_feed = vec![Feed {
-                group: "unique".to_string(),
-                name: "unique_feed".to_string(),
-                rss_link: "https://unique.example.com/different.xml".to_string(),
-            }];
-
-            let unique_result =
-                process_collect_article_links(&mock_client, &unique_feed, &pool).await;
-            assert!(
-                unique_result.is_ok(),
-                "ユニークフィードの処理が失敗しました"
-            );
-
-            // 新規フィードからの3件が追加されて、合計6件になることを確認
-            let final_unique_count = sqlx::query_scalar!("SELECT COUNT(*) FROM article_links")
-                .fetch_one(&pool)
-                .await?;
-            assert_eq!(
-                final_unique_count.unwrap_or(0),
-                6,
-                "新規フィード追加後は6件（既存3件 + 新規3件）になるべきです"
-            );
-
-            println!("✅ RSS重複処理テスト完了");
-            println!("  重複リンクは正しく上書き更新されました（日付が新しく更新）");
-            println!("  新規リンクは正しく追加されました（動的日付生成）");
-            println!("  最終リンク数: {}", final_unique_count.unwrap_or(0));
-
-            Ok(())
-        }
-    }
-
     // composition区分: 複数ドメインが横断するテスト（execute_rss_workflow統合テスト）
     mod composition {
         use super::*;
-        use crate::domain::feed::{search_feeds, FeedQuery};
+        use crate::core::feed::{search_feeds, FeedQuery};
         use crate::infra::api::{firecrawl::MockFirecrawlClient, http::MockHttpClient};
 
         /// 実際のfeeds.yamlを使用して、execute_rss_workflowが正しく動作することをテスト
