+++ src/core/article/model.rs
