+++ src/core/rss.rs
@@ -1,4 +1,4 @@
-use crate::domain::feed::Feed;
+use crate::core::feed::Feed;
 use crate::infra::api::http::HttpClient;
 use crate::infra::parser::{parse_channel_from_xml_str, parse_date};
 use anyhow::{Context, Result};
