+++ src/main.rs
@@ -1,9 +1,9 @@
 /// NOTE: main.rsは単なる最小限の動作確認に過ぎないので凝った実装をしないように
-use datadoggo::{app, domain, infra};
+use datadoggo::{app, core, infra};
 
 use app::workflow::execute_rss_workflow;
-use domain::feed::{search_feeds, FeedQuery};
-use domain::rss::{get_article_links_from_channel, store_article_links};
+use core::feed::{search_feeds, FeedQuery};
+use core::rss::{get_article_links_from_channel, store_article_links};
 use infra::api::firecrawl::ReqwestFirecrawlClient;
 use infra::api::http::ReqwestHttpClient;
 use infra::storage::db::setup_database;
