+++ src/task/mod.rs
@@ -1,0 +1,3 @@
+pub mod workflow;
+
+pub use workflow::{process_collect_article_links, process_collect_articles};
